+++ RELEASES.md
@@ -17,7 +17,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-metrics: added signature to metrics
 * validator: fixed issue with namespaces declaration in validation report
 * service-metrics: inserting (instead of overriding) type maps to maintain full history
-
+* service-metrics: type rating: traffic light for (lack of) attribution
 
 `2.2.0`
 * docker-service-templates: deployed dictionary for explanations, build for intel and arm platforms